Add permanent URLs
Every change in the form that allows uers to narrow down the type
of issues now creates a new URL using `history.pushState()`.

Each combination of filters creates a unique URL allowing to create
a permanent link for a specific configuration of filters.

Of course the results are not permanent as they depend on the state
of things in GitHub, but this allows users to link to specific filter
configurations, e.g. all starter issues.